requirements.txt
validators~=0.34.0
requests~=2.32.3
python-dotenv~=1.0.1
pymongo~=4.9.1
pathlib~=1.0.1
setuptools~=67.7.2